backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easy-listening-backend"
version = "0.1.0"
description = "Easy Listening backend service for real resource data"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.114,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "SQLAlchemy>=2.0.32,<3.0",
  "PyMySQL>=1.1.1,<2.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "httpx>=0.27,<1.0",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
